+++ pkg/config/config.go
@@ -1,7 +1,0 @@
-package config
-
-var (
-	Version   string
-	Commit    string
-	BuildTime string
-)