+++ tools/dependencies.sh
@@ -21,3 +21,6 @@ command -v go-winres &>/dev/null ||
 
 command -v cyclonedx-gomod &>/dev/null ||
   go install github.com/CycloneDX/cyclonedx-gomod/cmd/cyclonedx-gomod@latest
+
+command -v minify &>/dev/null ||
+  go install github.com/tdewolff/minify/v2@latest